Keep delegate and background-task runtime out of the coordinator core
The coordinator still owned a large delegate/background-task runtime shell that had nothing to do with the core coding turn loop. This extraction moves delegate execution, async/background enqueueing, and child-turn lifecycle handling into a dedicated delegate_runtime submodule so the coordinator root keeps converging toward the minimal runtime kernel.

Constraint: Preserve delegate, delegate_async, and background-task lifecycle behavior while reducing coordinator ownership
Rejected: Leave delegate/background-task runtime inside turn_coordinator.rs until a bigger cleanup | that would keep peripheral execution paths welded onto the core loop
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep delegate/background-task runtime helpers in delegate_runtime.rs and resist regrowing them inside the coordinator root
Tested: ./scripts/cargo-local-toolchain.sh check -p loong-app; ./scripts/cargo-local-toolchain.sh test -p loong-app spawn_background_delegate_with_runtime_creates_missing_root_session_scope -- --test-threads=1; ./scripts/cargo-local-toolchain.sh test -p loong-app spawn_background_delegate_with_runtime_uses_default_timeout_when_omitted -- --test-threads=1; ./scripts/cargo-local-toolchain.sh fmt --all -- --check; ./scripts/cargo-local-toolchain.sh clippy -p loong-app --all-targets -- -D warnings
Not-tested: full workspace test suite; the wider delegate_async lifecycle regression matrix in this slice
